Silence chktex "space before label" warnings
This is mostly to silence LaTeX's checker. Spaces before labels can in
some cases bause problems, but everything is fine here. This just tells
chktex to stop complaining about it.

While here, move a few labels to after the captions; this is necessary
because the caption creates the number that the label references, so
label must come after it.